Updates to model downloading/caching, project reorganization
