--- NeuralCodecs.Core/Loading/ModelLoadErrorEventArgs.cs
@@ -1,14 +1,0 @@
-namespace NeuralCodecs.Core.Loading
-{
-    public class ModelLoadErrorEventArgs : EventArgs
-    {
-        public string Source { get; }
-        public Exception Error { get; }
-
-        public ModelLoadErrorEventArgs(string source, Exception error)
-        {
-            Source = source;
-            Error = error;
-        }
-    }
-}
