--- NeuralCodecs.Core/Loading/ModelLoadWarningEventArgs.cs
@@ -1,12 +1,0 @@
-namespace NeuralCodecs.Core.Loading
-{
-    public class ModelLoadWarningEventArgs : EventArgs
-    {
-        public string Message { get; }
-
-        public ModelLoadWarningEventArgs(string message)
-        {
-            Message = message;
-        }
-    }
-}
