+++ NeuralCodecs.Core/Loading/ModelRegistry.cs
@@ -1,51 +1,40 @@
-﻿using Microsoft.Win32;
+﻿using NeuralCodecs.Core.Configuration;
 using NeuralCodecs.Core.Exceptions;
-using NeuralCodecs.Core.Interfaces;
-using NeuralCodecs.Core.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NeuralCodecs.Core.Loading
 {
+    /// <summary>
+    /// Registry for managing neural codec model factories and creation.
+    /// </summary>
     public class ModelRegistry
     {
-        //private readonly Dictionary<string, IModelFactory> _factories = new();
         private readonly Dictionary<Type, Func<IModelConfig, INeuralCodec>> _factories = new();
-        //public void RegisterFactory(string architecture, IModelFactory factory)
-        //{
-        //    _factories[architecture] = factory;
-        //}
 
-        //public INeuralCodec CreateModel(IModelConfig config)
-        //{
-        //    if (_factories.TryGetValue(config.Architecture, out var factory))
-        //    {
-        //        return factory.CreateModel(config);
-        //    }
-        //    throw new NotSupportedException($"No factory registered for {config.Architecture}");
-        //}
-        //public TModel CreateModel<TModel>(IModelConfig config) where TModel : class, INeuralCodec
-        //{
-        //    if (_factories.TryGetValue(config.Architecture, out var factory))
-        //    {
-        //        return factory.CreateModel(config) as TModel;
-        //    }
-        //    throw new NotSupportedException($"No factory registered for {config.Architecture}");
-        //}
+        /// <summary>
+        /// Registers a factory function for creating models of a specific type.
+        /// </summary>
+        /// <typeparam name="TModel">The type of model to register.</typeparam>
+        /// <typeparam name="TConfig">The type of configuration required by the model.</typeparam>
+        /// <param name="factory">The factory function that creates model instances.</param>
         public void RegisterModel<TModel, TConfig>(Func<TConfig, TModel> factory)
             where TModel : class, INeuralCodec
             where TConfig : class, IModelConfig
         {
-            //var configType = typeof(TConfig);
             var modelType = typeof(TModel);
 
             // Store the factory with type conversion
             _factories[modelType] = (config) => factory((TConfig)config);
         }
-        public TModel CreateModel<TModel, TConfig>(TConfig config) 
+
+        /// <summary>
+        /// Creates a model instance using the registered factory for the specified type.
+        /// </summary>
+        /// <typeparam name="TModel">The type of model to create.</typeparam>
+        /// <typeparam name="TConfig">The type of configuration required by the model.</typeparam>
+        /// <param name="config">The configuration to use for model creation.</param>
+        /// <returns>A new instance of the specified model type.</returns>
+        /// <exception cref="LoadException">Thrown when no factory is registered for the specified model type.</exception>
+        public TModel CreateModel<TModel, TConfig>(TConfig config)
             where TModel : class, INeuralCodec
             where TConfig : class, IModelConfig
         {
@@ -53,12 +42,11 @@ namespace NeuralCodecs.Core.Loading
 
             if (!_factories.TryGetValue(mType, out var factory))
             {
-                throw new ModelLoadException(
+                throw new LoadException(
                     $"No factory registered for type: {mType.Name}");
             }
 
             return (TModel)factory(config);
         }
-
     }
 }
