--- NeuralCodecs.Core/Models/CodecOperation.cs
@@ -1,10 +1,0 @@
-namespace NeuralCodecs.Core.Models
-{
-    public enum CodecOperation
-    {
-        Initialization,
-        Encoding,
-        Decoding,
-        ModelLoading
-    }
-}
