--- NeuralCodecs.Onnx/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace NeuralCodecs.Onnx
-{
-    public class Class1
-    {
-
-    }
-}
